Supporting modular form design
Separating out the input components from the form and using an array in the FormProps so you can list which inputs you need and the form will render them in that order. Currently only supports name (text), email, and message (long answer) inputs, should be expanded to include more types of data and allow for more than 3 different inputs
